+++ src/bittytax/price/bittytax_price.py
@@ -14,7 +14,7 @@ from colorama import Fore
 
 from ..bt_types import AssetSymbol, Timestamp
 from ..config import config
-from ..constants import ERROR, WARNING
+from ..constants import ERROR, TZ_UTC, WARNING
 from ..version import __version__
 from .assetdata import AsPriceRecord, AsRecord, AssetData
 from .datasource import DataSourceBase
@@ -305,7 +305,7 @@ def validate_date(value: str) -> Timestamp:
     except ValueError as e:
         raise argparse.ArgumentTypeError("date is not valid") from e
 
-    return Timestamp(date.replace(tzinfo=config.TZ_LOCAL))
+    return Timestamp(date.replace(tzinfo=TZ_UTC))
 
 
 def validate_quantity(value: str) -> Decimal: